refactor: Split create_table.rs tests into focused modules
Split the 947-line create_table.rs test file into modular structure:
- basic.rs: Basic CREATE TABLE tests (60 lines)
- data_types.rs: Type system tests (376 lines)
- constraints_column.rs: Column-level constraints (128 lines)
- constraints_table.rs: Table-level constraints (118 lines)
- string_types.rs: String type modifiers (189 lines)
- northwind.rs: Real-world schema examples (88 lines)

All 38 create_table tests pass successfully. Improves maintainability and
enables better test parallelization.

Note: Temporarily comments out datetime module in mod.rs due to pre-existing
build issue where datetime.rs is missing from main branch.

Closes #438
