+++ crates/parser/src/tests/create_table/mod.rs
@@ -1,0 +1,6 @@
+mod basic;
+mod constraints_column;
+mod constraints_table;
+mod data_types;
+mod northwind;
+mod string_types;
